fix(DiscountPercentage): error

+++ BeestjeOpJeFeestje/BeestjeOpJeFeestje/Controllers/BookingController.cs
@@ -197,7 +197,7 @@ namespace BeestjeOpJeFeestje.Controllers
             }
 
             double totalPrice = _priceRules.CalculateAnimalsPrice(animals);
-            (int discountPercentage, List<string> discountMessages) discountInfo = _priceRules.CalculateDiscount(animals, customerCard, DateTime.Parse(HttpContext.Session.GetString("SelectedDate")));
+            (double discountPercentage, List<string> discountMessages) discountInfo = _priceRules.CalculateDiscount(animals, customerCard, DateTime.Parse(HttpContext.Session.GetString("SelectedDate")));
 
             viewModel.TotalPrice = Math.Round(totalPrice * (1 - discountInfo.Item1 / 100), 2);
             viewModel.AppliedDiscounts = discountInfo.discountMessages;
